backend/requirements.txt
alembic==1.15.1
bcrypt==4.3.0
bleach==6.2.0
blinker==1.9.0
Brotli==1.1.0
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
Deprecated==1.2.18
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Compress==1.17
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.10.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
flask-talisman==1.1.0
Flask-Testing==0.8.1
Flask-WTF==1.2.2
gunicorn==23.0.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
limits==4.1
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
ordered-set==4.1.0
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
pyOpenSSL==25.0.0
pytest==8.3.5
python-dotenv==1.0.1
rich==13.9.4
SQLAlchemy==2.0.38
typing_extensions==4.12.2
webencodings==0.5.1
Werkzeug==3.1.3
wrapt==1.17.2
WTForms==3.2.1
zstandard==0.23.0
